# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Group Chat API",
    "description": "API for managing group chats and messages",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "tags": [
    {
      "name": "Authentication",
      "description": "Endpoints for logging in and out"
    },
    {
      "name": "Users",
      "description": "Endpoints for managing users"
    },
    {
      "name": "Groups",
      "description": "Endpoints for managing groups"
    },
    {
      "name": "Messages",
      "description": "Endpoints for managing messages"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logs a user in and returns an access token",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/LoginUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "UNAUTHORIZED"
          }
        }
      }
    },
    "/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logs a user out and invalidates their access token"
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Returns a list of all users",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "User list fetched",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "summary": "Create a new user",
        "tags": [
          "Users"
        ],
        "operationId": "create_user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/CreateUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "user_id needs to be fetched",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "404": {
            "description": "User not found"
          },
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        },
        "summary": "Returns information about a specific user",
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "ID user_id needs to be fetched",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/CreateUser"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "User not found"
          },
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        },
        "summary": "Updates information about a specific user",
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "user_id needs to be fetched",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "404": {
            "description": "User not found"
          },
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        },
        "summary": "Deletes a specific user",
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/groups": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Returns a list of all groups",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "group prefix that need to be considered for search",
            "required": false,
            "type": "string",
            "collectionFormat": "single"
          }
        ],
        "responses": {
          "200": {
            "description": "Group list fetched",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Group"
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "Groups"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/CreateGroup"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Group created"
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "summary": "Creates a new group"
      }
    },
    "/groups/{group_id}": {
      "get": {
        "tags": [
          "Groups"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "group id that need to be considered for search",
            "required": false,
            "type": "int64",
            "collectionFormat": "single"
          }
        ],
        "responses": {
          "200": {
            "description": "Group fetched",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Group"
            }
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "summary": "Returns information about a specific group",
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Groups"
        ],
        "summary": "Deletes a specific group",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "group id that need to be considered to delete",
            "required": true,
            "type": "integer",
            "format": "int64",
            "collectionFormat": "single"
          }
        ],
        "security": [
          {
            "api_key": []
          }
        ],
        "responses": {
          "200": {
            "description": "Group deleted"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/groups/{group_id}/members": {
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Adds a user to a specific group",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "group id that need to be considered to add user",
            "required": true,
            "type": "integer",
            "format": "int64",
            "collectionFormat": "single"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GroupMember"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "user added"
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/groups/{group_id}/messages/{message_id}/like": {
      "post": {
        "tags": [
          "Messages"
        ], "summary": "like s messages in a specific group",
        "produces": [
          "application/json"
        ], "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "group id that need to be considered to like messages",
            "required": true,
            "type": "integer",
            "format": "int64",
            "collectionFormat": "single"
          },{
            "name": "message_id",
            "in": "path",
            "description": "message_id  that need to be considered to like message",
            "required": true,
            "type": "integer",
            "format": "int64",
            "collectionFormat": "single"
          }
        ],"responses": {
          "200": {
            "description": "message liked"
          },
          "400": {
            "description": "Invalid input"
          }
        },"security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/groups/{group_id}/messages": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Returns a list of all messages in a specific group",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "group id that need to be considered to fetch messages",
            "required": true,
            "type": "integer",
            "format": "int64",
            "collectionFormat": "single"
          }
        ],
        "responses": {
          "200": {
            "description": "user added",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Message"
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "Adds a new message to a specific group",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "group id that need to be considered to send messages",
            "required": true,
            "type": "integer",
            "format": "int64",
            "collectionFormat": "single"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MessageText"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "message sent"
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "username": {
          "type": "string"
        },
        "is_admin": {
          "type": "boolean",
          "default": false
        }
      }
    },
    "LoginUser": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "CreateUser": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "is_admin": {
          "type": "boolean",
          "default": false
        }
      }
    },
    "Group": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "GroupMember": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        }
      }
    },
    "CreateGroup": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "Message": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "user_id": {
          "type": "integer"
        },
        "group_id": {
          "type": "integer"
        },
        "text": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "likes": {
          "type": "array",
          "items": {
            "type": "integer"
          }
        }
      }
    },
    "MessageText": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string"
        }
      }
    }
  }
}